I moved most of the infrastructure for this page from studies to proposals, cause I thought that would make more sense. Als got the 'return to start of apllication' button working.
