+++ proposals/views/study_views.py
@@ -7,7 +7,7 @@ from main.views import AllowErrorsOnBackbuttonMixin, UpdateView, FormSetUpdateVi
 from studies.models import Documents, Study
 from studies.forms import StudyConsentForm
 from studies.utils import create_documents_for_study
-from ..forms import StudyStartForm, TranslatedConsentForms
+from ..forms import StudyStartForm
 from ..models import Proposal
 
 
@@ -112,29 +112,9 @@ class StudyConsent(AllowErrorsOnBackbuttonMixin, FormSetUpdateView):
         Otherwise, go to the data management view.
         """
         proposal = Proposal.objects.get(pk=self.kwargs.get('pk'))
-        return reverse('studies:translated', args=(proposal.pk,))
+        return reverse('proposals:translated', args=(proposal.pk,))
 
     def get_back_url(self):
         """Return to the Study design view"""
         proposal = Proposal.objects.get(pk=self.kwargs.get('pk'))
         return reverse('studies:design_end', args=(proposal.last_study().pk,))
-
-class TranslatedConsentFormsView(UpdateView):
-    model = Proposal
-    form_class = TranslatedConsentForms
-    template_name = 'proposals/translated_consent_forms.html'
-
-    # def get_context_data(self, *args, **kwargs):
-    #     """Setting the progress on the context"""  # (??)
-    #     context = super().get_context_data(*args, **kwargs)
-    #     # The following is used by the progress bar
-    #     proposal = Proposal.objects.get(pk=self.kwargs.get('pk'))
-    #     context['proposal'] = proposal
-
-    def get_next_url(self):
-        proposal = Proposal.objects.get(pk=self.kwargs.get('pk'))
-        return reverse('proposals:data_management', args=(proposal.pk,))
-
-    def get_back_url(self):
-        """Return to the consent form overview of the last Study"""
-        return reverse('proposals:consent', args=(self.object.pk,))
